move TRPCException class to separate files
