+++ src/rpc/rpc_exception.h
@@ -1,0 +1,28 @@
+#include <stdexcept>
+#include <string>
+
+// RPC Request execution result code
+enum class TRPCResultCode
+{
+    // No errors
+    RPC_OK = 0,
+    // Wrong parameters value
+    RPC_WRONG_PARAM_VALUE = -1,
+    // Requested port was not found
+    RPC_WRONG_PORT = -2,
+    // Unsuccessful port IO
+    RPC_WRONG_IO = -3,
+    // RPC request handling timeout
+    RPC_WRONG_TIMEOUT = -4
+};
+
+class TRPCException: public std::runtime_error
+{
+public:
+    TRPCException(const std::string& message, TRPCResultCode resultCode);
+    TRPCResultCode GetResultCode() const;
+    std::string GetResultMessage() const;
+
+private:
+    TRPCResultCode ResultCode;
+};
