+++ src/rpc/rpc_port_driver_list.h
@@ -1,24 +1,10 @@
 #pragma once
 #include "rpc_config.h"
+#include "rpc_exception.h"
 #include "serial_driver.h"
 #include "serial_port_driver.h"
 #include <wblib/json_utils.h>
 #include <wblib/rpc.h>
-
-// RPC Request execution result code
-enum class TRPCResultCode
-{
-    // No errors
-    RPC_OK = 0,
-    // Wrong parameters value
-    RPC_WRONG_PARAM_VALUE = -1,
-    // Requested port was not found
-    RPC_WRONG_PORT = -2,
-    // Unsuccessful port IO
-    RPC_WRONG_IO = -3,
-    // RPC request handling timeout
-    RPC_WRONG_TIMEOUT = -4
-};
 
 struct TRPCPortDriver
 {
@@ -27,17 +13,6 @@ struct TRPCPortDriver
 };
 
 typedef std::shared_ptr<TRPCPortDriver> PRPCPortDriver;
-
-class TRPCException: public std::runtime_error
-{
-public:
-    TRPCException(const std::string& message, TRPCResultCode resultCode);
-    TRPCResultCode GetResultCode() const;
-    std::string GetResultMessage() const;
-
-private:
-    TRPCResultCode ResultCode;
-};
 
 class TRPCPortDriverList
 {
